+++ src/main/java/fr/insee/genesis/domain/ports/api/EditedResponseApiPort.java
@@ -1,7 +1,0 @@
-package fr.insee.genesis.domain.ports.api;
-
-import fr.insee.genesis.domain.model.editedresponse.EditedResponseModel;
-
-public interface EditedResponseApiPort {
-    EditedResponseModel getEditedResponse(String questionnaireId, String interrogationId);
-}