+++ src/test/java/cucumber/functional_tests/config/CucumberSpringConfiguration.java
@@ -1,9 +1,9 @@
 package cucumber.functional_tests.config;
 
 import fr.insee.genesis.GenesisApi;
+import fr.insee.genesis.infrastructure.repository.ContextualExternalVariableMongoDBRepository;
+import fr.insee.genesis.infrastructure.repository.ContextualPreviousVariableMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.DataProcessingContextMongoDBRepository;
-import fr.insee.genesis.infrastructure.repository.EditedExternalResponseMongoDBRepository;
-import fr.insee.genesis.infrastructure.repository.EditedPreviousResponseMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.LastJsonExtractionMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.LunaticJsonMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.LunaticModelMongoDBRepository;
@@ -42,9 +42,9 @@ public class CucumberSpringConfiguration {
     @MockitoBean
     private DataProcessingContextMongoDBRepository dataProcessingContextMongoDBRepository;
     @MockitoBean
-    private EditedPreviousResponseMongoDBRepository editedPreviousResponseMongoDBRepository;
+    private ContextualPreviousVariableMongoDBRepository contextualPreviousVariableMongoDBRepository;
     @MockitoBean
-    private EditedExternalResponseMongoDBRepository editedExternalResponseMongoDBRepository;
+    private ContextualExternalVariableMongoDBRepository contextualExternalVariableMongoDBRepository;
     @MockitoBean
     private QuestionnaireMetadataMongoDBRepository questionnaireMetadataMongoDBRepository;
 }