+++ src/test/java/fr/insee/genesis/controller/rest/ControllerAccessTest.java
@@ -3,9 +3,9 @@ package fr.insee.genesis.controller.rest;
 import fr.insee.genesis.domain.ports.api.DataProcessingContextApiPort;
 import fr.insee.genesis.domain.ports.api.LunaticJsonRawDataApiPort;
 import fr.insee.genesis.domain.ports.api.SurveyUnitApiPort;
+import fr.insee.genesis.infrastructure.repository.ContextualExternalVariableMongoDBRepository;
+import fr.insee.genesis.infrastructure.repository.ContextualPreviousVariableMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.DataProcessingContextMongoDBRepository;
-import fr.insee.genesis.infrastructure.repository.EditedExternalResponseMongoDBRepository;
-import fr.insee.genesis.infrastructure.repository.EditedPreviousResponseMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.LastJsonExtractionMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.LunaticJsonMongoDBRepository;
 import fr.insee.genesis.infrastructure.repository.LunaticModelMongoDBRepository;
@@ -91,9 +91,9 @@ class ControllerAccessTest {
     @MockitoBean
     private LunaticModelMongoDBRepository lunaticModelMongoDBRepository;
     @MockitoBean
-    private EditedPreviousResponseMongoDBRepository editedPreviousResponseMongoDBRepository;
+    private ContextualPreviousVariableMongoDBRepository contextualPreviousVariableMongoDBRepository;
     @MockitoBean
-    private EditedExternalResponseMongoDBRepository editedExternalResponseMongoDBRepository;
+    private ContextualExternalVariableMongoDBRepository contextualExternalVariableMongoDBRepository;
 
     @MockitoBean
     private QuestionnaireMetadataMongoDBRepository questionnaireMetadataMongoDBRepository;