+++ src/test/java/fr/insee/genesis/controller/rest/responses/QuestionnaireControllerTest.java
@@ -2,10 +2,13 @@ package fr.insee.genesis.controller.rest.responses;
 
 import fr.insee.genesis.controller.dto.QuestionnaireWithCampaign;
 import fr.insee.genesis.domain.ports.api.SurveyUnitApiPort;
+import fr.insee.genesis.domain.service.context.DataProcessingContextService;
 import fr.insee.genesis.domain.service.metadata.QuestionnaireMetadataService;
 import fr.insee.genesis.domain.service.surveyunit.SurveyUnitService;
+import fr.insee.genesis.infrastructure.document.context.DataProcessingContextDocument;
 import fr.insee.genesis.infrastructure.utils.FileUtils;
 import fr.insee.genesis.stubs.ConfigStub;
+import fr.insee.genesis.stubs.DataProcessingContextPersistancePortStub;
 import fr.insee.genesis.stubs.QuestionnaireMetadataPersistancePortStub;
 import fr.insee.genesis.stubs.SurveyUnitPersistencePortStub;
 import org.assertj.core.api.Assertions;
@@ -15,6 +18,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.ResponseEntity;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
@@ -24,23 +28,29 @@ class QuestionnaireControllerTest {
     //Given
     static QuestionnaireController questionnaireControllerStatic;
     static SurveyUnitPersistencePortStub surveyUnitPersistencePortStub;
+    static DataProcessingContextPersistancePortStub dataProcessingContextPersistancePortStub;
 
     @BeforeAll
     static void init() {
         surveyUnitPersistencePortStub = new SurveyUnitPersistencePortStub();
+        dataProcessingContextPersistancePortStub = new DataProcessingContextPersistancePortStub();
         SurveyUnitApiPort surveyUnitApiPort = new SurveyUnitService(
                 surveyUnitPersistencePortStub,
                 new QuestionnaireMetadataService(new QuestionnaireMetadataPersistancePortStub()),
                 new FileUtils(new ConfigStub())
         );
 
-        questionnaireControllerStatic = new QuestionnaireController(surveyUnitApiPort);
+        questionnaireControllerStatic = new QuestionnaireController(
+                surveyUnitApiPort,
+                new DataProcessingContextService(dataProcessingContextPersistancePortStub, surveyUnitPersistencePortStub)
+        );
 
     }
 
     @BeforeEach
     void reset() throws IOException {
-       Utils.reset(surveyUnitPersistencePortStub);
+        dataProcessingContextPersistancePortStub.getMongoStub().clear();
+        Utils.reset(surveyUnitPersistencePortStub);
     }
 
 
@@ -55,6 +65,79 @@ class QuestionnaireControllerTest {
         Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
         Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsOnly(
                 DEFAULT_QUESTIONNAIRE_ID,"TESTQUESTIONNAIRE2");
+    }
+
+    @Test
+    void getQuestionnairesWithReviewTest() {
+
+        String questionnaireId = "TESTQUESTIONNAIRE2";
+        Utils.addAdditionalSurveyUnitModelToMongoStub(questionnaireId, surveyUnitPersistencePortStub);
+
+        ResponseEntity<List<String>> response = questionnaireControllerStatic.getQuestionnairesWithReview(true);
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isEmpty();
+
+        dataProcessingContextPersistancePortStub.getMongoStub().add(
+                new DataProcessingContextDocument(
+                        DEFAULT_QUESTIONNAIRE_ID,
+                        new ArrayList<>(),
+                        true
+                )
+        );
+
+        response = questionnaireControllerStatic.getQuestionnairesWithReview(true);
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsOnly(
+                DEFAULT_QUESTIONNAIRE_ID);
+
+        dataProcessingContextPersistancePortStub.getMongoStub().add(
+                new DataProcessingContextDocument(
+                        questionnaireId,
+                        new ArrayList<>(),
+                        false
+                )
+        );
+
+        response = questionnaireControllerStatic.getQuestionnairesWithReview(true);
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsOnly(
+                DEFAULT_QUESTIONNAIRE_ID);
+    }
+
+    @Test
+    void getQuestionnairesWithoutReviewTest() {
+
+        String questionnaireId = "TESTQUESTIONNAIRE2";
+        Utils.addAdditionalSurveyUnitModelToMongoStub(questionnaireId, surveyUnitPersistencePortStub);
+
+        ResponseEntity<List<String>> response = questionnaireControllerStatic.getQuestionnairesWithReview(false);
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isEmpty();
+
+        dataProcessingContextPersistancePortStub.getMongoStub().add(
+                new DataProcessingContextDocument(
+                        DEFAULT_QUESTIONNAIRE_ID,
+                        new ArrayList<>(),
+                        false
+                )
+        );
+
+        response = questionnaireControllerStatic.getQuestionnairesWithReview(false);
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsOnly(
+                DEFAULT_QUESTIONNAIRE_ID);
+
+        dataProcessingContextPersistancePortStub.getMongoStub().add(
+                new DataProcessingContextDocument(
+                        questionnaireId,
+                        new ArrayList<>(),
+                        true
+                )
+        );
+        response = questionnaireControllerStatic.getQuestionnairesWithReview(false);
+        Assertions.assertThat(response.getStatusCode().is2xxSuccessful()).isTrue();
+        Assertions.assertThat(response.getBody()).isNotNull().isNotEmpty().containsOnly(
+                DEFAULT_QUESTIONNAIRE_ID);
     }
 
     @Test